src/backend copy/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-rest-swagger
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
